Refactor: New metadata service
