+++ src/app/app.module.ts
@@ -100,6 +100,7 @@ import {
   GroupService,
   ImageGroupService,
   LogService,
+  MetadataService,
   TitleService,
   ToolboxService,
   TypeIdPipe,
@@ -139,6 +140,7 @@ const APP_PROVIDERS = [
   LegacyRouteResolver,
   Title,
   TitleService,
+  MetadataService,
   { provide: UrlSerializer, useClass: CustomUrlSerializer },
   { provide: HTTP_INTERCEPTORS, useClass: UnauthorizedInterceptor, multi: true }
   // { provide: RouteReuseStrategy, useClass: CustomReuseStrategy } // to be implemented later
