+++ src/app/shared/index.ts
@@ -4,6 +4,7 @@ export * from './auth.service'
 export * from './assets.service'
 export * from './asset-search.service'
 export * from './group.service'
+export * from './metadata.service'
 export * from './toolbox.service'
 export * from './nav'
 export * from './footer'
